Move legacy wrapper into original module
as requested by @cdent, despite my reservations (cf. rationale in
previous commit)

note that this function is side-effecty WRT deprecation warnings, which
is necessary as Python 2.7+ silences them by default (according to
https://code.djangoproject.com/ticket/18985)
